move common_parameters out of private
private is not installed so should only be used for implementation files
