+++ include/adm/elements/audio_block_format_direct_speakers.hpp
@@ -5,7 +5,7 @@
 #include <boost/variant.hpp>
 #include "adm/elements/time.hpp"
 #include "adm/elements/audio_block_format_id.hpp"
-#include "adm/elements/private/common_parameters.hpp"
+#include "adm/elements/common_parameters.hpp"
 #include "adm/elements/speaker_position.hpp"
 #include "adm/elements_fwd.hpp"
 #include "adm/detail/named_option_helper.hpp"
