+++ include/adm/elements/audio_block_format_matrix.hpp
@@ -4,7 +4,7 @@
 #include <boost/optional.hpp>
 #include "adm/elements/time.hpp"
 #include "adm/elements/audio_block_format_id.hpp"
-#include "adm/elements/private/common_parameters.hpp"
+#include "adm/elements/common_parameters.hpp"
 #include "adm/elements_fwd.hpp"
 #include "adm/detail/named_option_helper.hpp"
 #include "adm/detail/named_type.hpp"
